+++ Projects/App/Project.swift
@@ -14,7 +14,8 @@ let project = Project.makeModule(
             path: .relativeToRoot("Projects/Core/JwtStore")),
         TargetDependency.project(
             target: "NetworksMonitor",
-            path: .relativeToRoot("Projects/Core/NetworksMonitor"))
+            path: .relativeToRoot("Projects/Core/NetworksMonitor")),
+        .SPM.GoogleMobileAds
     ],
     resources: ["Resources/**"],
     infoPlist: .file(path: "Support/Info.plist")
