+++ Projects/App/Support/Info.plist
@@ -2,6 +2,12 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>NSUserTrackingUsageDescription</key>
+	<string>This is identifier..</string>
+	<key>GADApplicationIdentifier</key>
+	<string>ca-app-pub-5088279003431597~8818213465</string>
+	<key>GADIsAdManagerApp</key>
+	<string>1</string>
 	<key>CFBundleDevelopmentRegion</key>
 	<string>$(DEVELOPMENT_LANGUAGE)</string>
 	<key>CFBundleExecutable</key>
